Rotation based on Edge detection
In this code, the robot will follow the inner wall solely based on its color. This will serve as the basis for creating a program that enables the robot to adapt to changes in the inner walling in the real game. We will look further into how to do this perfectly with color sensors only and the scarcity of ultrasonic sensors.