Edge Detection using Ultrasonic Sensors
We found that using color sensors are inefficient for the wall detector. With EV3 ultrasonic sensors, the robot efficiently running on the map with no obstacles in the front. This will be our basis for the full run with all sensors intact, obstacles placed, and adaptable in the different game scenarios.